% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_Color_TestMatrix_Ncol_10_Nrow_16.R
\docType{data}
\name{Color_TestMatrix_Ncol_10_Nrow_16}
\alias{Color_TestMatrix_Ncol_10_Nrow_16}
\title{Color_TestMatrix_Ncol_10_Nrow_16.RData}
\format{An object of class \code{matrix} with 16 rows and 10 columns.}
\usage{
data(Color_TestMatrix_Ncol_10_Nrow_16)
}
\description{
Example dataset with 10  columns and 16 rows which test default color ranking.
}
\examples{
Produce decision table with same dimensions as example dataset. Save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = Color_TestMatrix_Ncol_10_Nrow_16, BestPerformanceVector = c("High","High","High","High","High","High","High","High","High","High"))

}
\keyword{datasets}
