% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_Color_TestMatrix_Ncol_11_Nrow_10.R
\docType{data}
\name{Color_TestMatrix_Ncol_11_Nrow_10}
\alias{Color_TestMatrix_Ncol_11_Nrow_10}
\title{Color_TestMatrix_Ncol_11_Nrow_10.RData}
\format{An object of class \code{matrix} with 10 rows and 11 columns.}
\usage{
data(Color_TestMatrix_Ncol_11_Nrow_10)
}
\description{
Example dataset with 11  columns and 10 rows which test default color ranking.
}
\examples{
Produce decision table with same dimensions as example dataset. Save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = Color_TestMatrix_Ncol_11_Nrow_10, BestPerformanceVector = c("High","High","High","High","High","High","High","High","High","High","High"))

}
\keyword{datasets}
