% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_Color_TestMatrix_Ncol_18_Nrow_18.R
\docType{data}
\name{Color_TestMatrix_Ncol_18_Nrow_18}
\alias{Color_TestMatrix_Ncol_18_Nrow_18}
\title{Color_TestMatrix_Ncol_18_Nrow_18.RData}
\format{An object of class \code{matrix} with 18 rows and 18 columns.}
\usage{
data(Color_TestMatrix_Ncol_18_Nrow_18)
}
\description{
Example dataset with 18  columns and 18 rows which test default color ranking.
}
\examples{
Produce decision table with same dimensions as example dataset. Save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = Color_TestMatrix_Ncol_18_Nrow_18, BestPerformanceVector = c("High","High","High","High","High","High","High","High","High","High","High","High","High","High","High","High","High","High"))

}
\keyword{datasets}
