% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_Color_TestMatrix_Ncol_1_Nrow_3.R
\docType{data}
\name{Color_TestMatrix_Ncol_1_Nrow_3}
\alias{Color_TestMatrix_Ncol_1_Nrow_3}
\title{Color_TestMatrix_Ncol_1_Nrow_3.RData}
\format{An object of class \code{matrix} with 3 rows and 1 columns.}
\usage{
data(Color_TestMatrix_Ncol_1_Nrow_3)
}
\description{
Example dataset with 1  columns and 3 rows which test default color ranking.
}
\examples{
Produce decision table with same dimensions as example dataset. Save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = Color_TestMatrix_Ncol_1_Nrow_3, BestPerformanceVector = c("High","High","High"))

}
\keyword{datasets}
