% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_Color_TestMatrix_Ncol_1_Nrow_8.R
\docType{data}
\name{Color_TestMatrix_Ncol_1_Nrow_8}
\alias{Color_TestMatrix_Ncol_1_Nrow_8}
\title{Color_TestMatrix_Ncol_1_Nrow_8.RData}
\format{An object of class \code{matrix} with 8 rows and 1 columns.}
\usage{
data(Color_TestMatrix_Ncol_1_Nrow_8)
}
\description{
Example dataset with 1  columns and 8 rows which test default color ranking.
}
\examples{
Produce decision table with same dimensions as example dataset. Save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = Color_TestMatrix_Ncol_1_Nrow_8, BestPerformanceVector = c("High","High","High"))

}
\keyword{datasets}
