% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_Color_TestMatrix_Ncol_20_Nrow_13.R
\docType{data}
\name{Color_TestMatrix_Ncol_20_Nrow_13}
\alias{Color_TestMatrix_Ncol_20_Nrow_13}
\title{Color_TestMatrix_Ncol_20_Nrow_13.RData}
\format{An object of class \code{matrix} with 13 rows and 20 columns.}
\usage{
data(Color_TestMatrix_Ncol_20_Nrow_13)
}
\description{
Example dataset with 20  columns and 13 rows which test default color ranking.
}
\examples{
Produce decision table with same dimensions as example dataset. Save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = Color_TestMatrix_Ncol_20_Nrow_13, BestPerformanceVector = c("High","High","High","High","High","High","High","High","High","High","High","High","High","High","High","High","High","High","High","High"))

}
\keyword{datasets}
