% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_Color_TestMatrix_Ncol_2_Nrow_6.R
\docType{data}
\name{Color_TestMatrix_Ncol_2_Nrow_6}
\alias{Color_TestMatrix_Ncol_2_Nrow_6}
\title{Color_TestMatrix_Ncol_2_Nrow_6.RData}
\format{An object of class \code{matrix} with 6 rows and 2 columns.}
\usage{
data(Color_TestMatrix_Ncol_2_Nrow_6)
}
\description{
Example dataset with 2  columns and 6 rows which test default color ranking.
}
\examples{
Produce decision table with same dimensions as example dataset. Save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = Color_TestMatrix_Ncol_2_Nrow_6, BestPerformanceVector = c("High","High"))

}
\keyword{datasets}
