% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_Color_TestMatrix_Ncol_4_Nrow_7.R
\docType{data}
\name{Color_TestMatrix_Ncol_4_Nrow_7}
\alias{Color_TestMatrix_Ncol_4_Nrow_7}
\title{Color_TestMatrix_Ncol_4_Nrow_7.RData}
\format{An object of class \code{matrix} with 7 rows and 4 columns.}
\usage{
data(Color_TestMatrix_Ncol_4_Nrow_7)
}
\description{
Example dataset with 4  columns and 7 rows which test default color ranking.
}
\examples{
Produce decision table with same dimensions as example dataset. Save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = Color_TestMatrix_Ncol_4_Nrow_7, BestPerformanceVector = c("High","High","High","High"))

}
\keyword{datasets}
