% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_Color_TestMatrix_Ncol_5_Nrow_20.R
\docType{data}
\name{Color_TestMatrix_Ncol_5_Nrow_20}
\alias{Color_TestMatrix_Ncol_5_Nrow_20}
\title{Color_TestMatrix_Ncol_5_Nrow_20.RData}
\format{An object of class \code{matrix} with 20 rows and 5 columns.}
\usage{
data(Color_TestMatrix_Ncol_5_Nrow_20)
}
\description{
Example dataset with 5  columns and 20 rows which test default color ranking.
}
\examples{
Produce decision table with same dimensions as example dataset. Save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = Color_TestMatrix_Ncol_5_Nrow_20, BestPerformanceVector = c("High","High","High","High","High"))

}
\keyword{datasets}
