% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_Color_TestMatrix_Ncol_6_Nrow_19.R
\docType{data}
\name{Color_TestMatrix_Ncol_6_Nrow_19}
\alias{Color_TestMatrix_Ncol_6_Nrow_19}
\title{Color_TestMatrix_Ncol_6_Nrow_19.RData}
\format{An object of class \code{matrix} with 19 rows and 6 columns.}
\usage{
data(Color_TestMatrix_Ncol_6_Nrow_19)
}
\description{
Example dataset with 6  columns and 19 rows which test default color ranking.
}
\examples{
Produce decision table with same dimensions as example dataset. Save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = Color_TestMatrix_Ncol_6_Nrow_19, BestPerformanceVector = c("High","High","High","High","High","High"))

}
\keyword{datasets}
