% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_Color_TestMatrix_Ncol_7_Nrow_4.R
\docType{data}
\name{Color_TestMatrix_Ncol_7_Nrow_4}
\alias{Color_TestMatrix_Ncol_7_Nrow_4}
\title{Color_TestMatrix_Ncol_7_Nrow_4.RData}
\format{An object of class \code{matrix} with 4 rows and 7 columns.}
\usage{
data(Color_TestMatrix_Ncol_7_Nrow_4)
}
\description{
Example dataset with 7  columns and 4 rows which test default color ranking.
}
\examples{
Produce decision table with same dimensions as example dataset. Save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = Color_TestMatrix_Ncol_7_Nrow_4, BestPerformanceVector = c("High","High","High","High","High","High","High"))

}
\keyword{datasets}
