% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_Color_TestMatrix_Ncol_8_Nrow_17.R
\docType{data}
\name{Color_TestMatrix_Ncol_8_Nrow_17}
\alias{Color_TestMatrix_Ncol_8_Nrow_17}
\title{Color_TestMatrix_Ncol_8_Nrow_17.RData}
\format{An object of class \code{matrix} with 17 rows and 8 columns.}
\usage{
data(Color_TestMatrix_Ncol_8_Nrow_17)
}
\description{
Example dataset with 8  columns and 17 rows which test default color ranking.
}
\examples{
Produce decision table with same dimensions as example dataset. Save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = Color_TestMatrix_Ncol_8_Nrow_17, BestPerformanceVector = c("High","High","High","High","High","High","High","High"))

}
\keyword{datasets}
