% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_Color_TestMatrix_Ncol_9_Nrow_1.R
\docType{data}
\name{Color_TestMatrix_Ncol_9_Nrow_1}
\alias{Color_TestMatrix_Ncol_9_Nrow_1}
\title{Color_TestMatrix_Ncol_9_Nrow_1.RData}
\format{An object of class \code{matrix} with 1 rows and 9 columns.}
\usage{
data(Color_TestMatrix_Ncol_9_Nrow_1)
}
\description{
Example dataset with 9  columns and 1 rows which test default color ranking.
}
\examples{
Produce decision table with same dimensions as example dataset. Save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = Color_TestMatrix_Ncol_9_Nrow_1, BestPerformanceVector = c("High","High","High","High","High","High","High","High","High"))

}
\keyword{datasets}
