% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_Color_TestMatrix_Ncol_9_Nrow_12.R
\docType{data}
\name{Color_TestMatrix_Ncol_9_Nrow_12}
\alias{Color_TestMatrix_Ncol_9_Nrow_12}
\title{Color_TestMatrix_Ncol_9_Nrow_12.RData}
\format{An object of class \code{matrix} with 12 rows and 9 columns.}
\usage{
data(Color_TestMatrix_Ncol_9_Nrow_12)
}
\description{
Example dataset with 9  columns and 12 rows which test default color ranking.
}
\examples{
Produce decision table with same dimensions as example dataset. Save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = Color_TestMatrix_Ncol_9_Nrow_12, BestPerformanceVector = c("High","High","High","High","High","High","High","High","High"))

}
\keyword{datasets}
