% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8Col_Graphics.R
\name{GraphicFormat8Col}
\alias{GraphicFormat8Col}
\title{Select format for decision tables with 8 columns.}
\usage{
GraphicFormat8Col(Data, SummaryRowOption = "Off")
}
\arguments{
\item{Data}{A matrix with row and column names, no default}

\item{SummaryRowOption}{A string indicating summary row option, default = no summary row
"Off"          No summary row
"MeanRank"     Summary row added, values represent mean rank across all rows in each column 
"SumRank"      Summary row added, values represent summed rank across all rows in each column}
}
\value{
GraphicFormat$GraphicNRow        Number of rows in graphic layout
             $GraphicNCol        Number of columns in graphic layout
             $GraphicLaoyout     Structure of graphic layout
             $GraphicBarColors   List of default colors for graphic bars# Format for Decision tables with 2 columns
}
\description{
This function selects the format for decision tables with 8 columns and between 1 and 20 rows.
}
