% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_RData.R
\docType{data}
\name{Icon_Environmental_Considerations}
\alias{Icon_Environmental_Considerations}
\title{Raster data to display the Environmental Considerations Icon}
\format{An object of class \code{SpatialPixelsDataFrame} with 1788600 rows and 1 columns.}
\usage{
Icon_Environmental_Considerations
}
\description{
Raster data to display the Environmental Considerations Icon
}
\keyword{datasets}
