% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_RData.R
\docType{data}
\name{Icon_Groundfish_Fishery}
\alias{Icon_Groundfish_Fishery}
\title{Raster data to display the Groundfish Fishery Icon}
\format{
An object of class \code{SpatialPixelsDataFrame} with 16383 rows and 1 columns.
}
\usage{
Icon_Groundfish_Fishery
}
\description{
Raster data to display the Groundfish Fishery Icon
}
\keyword{datasets}
