% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_RData.R
\docType{data}
\name{Icon_Groundfish_Species}
\alias{Icon_Groundfish_Species}
\title{Raster data to display the Groundfish Species Icon}
\format{An object of class \code{SpatialPixelsDataFrame} with 1636988 rows and 1 columns.}
\usage{
Icon_Groundfish_Species
}
\description{
Raster data to display the Groundfish Species Icon
}
\keyword{datasets}
