% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_RData.R
\docType{data}
\name{Icon_Herring_Fishery_Option2}
\alias{Icon_Herring_Fishery_Option2}
\title{Raster data to display the Herring Fishery Icon Option 2}
\format{
An object of class \code{SpatialPixelsDataFrame} with 17415 rows and 1 columns.
}
\usage{
Icon_Herring_Fishery_Option2
}
\description{
Raster data to display the Herring Fishery Icon Option 2
}
\keyword{datasets}
