% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_RData.R
\docType{data}
\name{Icon_Lobster}
\alias{Icon_Lobster}
\title{Raster data to display the Lobster Icon}
\format{
An object of class \code{SpatialPixelsDataFrame} with 17030 rows and 1 columns.
}
\usage{
Icon_Lobster
}
\description{
Raster data to display the Lobster Icon
}
\keyword{datasets}
