% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_RData.R
\docType{data}
\name{Icon_Lobster_Fishery}
\alias{Icon_Lobster_Fishery}
\title{Raster data to display the Lobster Fishery Icon}
\format{
An object of class \code{SpatialPixelsDataFrame} with 16254 rows and 1 columns.
}
\usage{
Icon_Lobster_Fishery
}
\description{
Raster data to display the Lobster Fishery Icon
}
\keyword{datasets}
