% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_RData.R
\docType{data}
\name{Icon_Predator_Species_Tuna_Haddock_Flatfish}
\alias{Icon_Predator_Species_Tuna_Haddock_Flatfish}
\title{Raster data to display the Predator Species Icon for Tuna, Haddock, and Flatfish}
\format{An object of class \code{SpatialPixelsDataFrame} with 1768275 rows and 1 columns.}
\usage{
Icon_Predator_Species_Tuna_Haddock_Flatfish
}
\description{
Raster data to display the Predator Species Icon for Tuna, Haddock, and Flatfish
}
\keyword{datasets}
