% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_RData.R
\docType{data}
\name{Icon_Primary_Production}
\alias{Icon_Primary_Production}
\title{Raster data to display the Primary Production Icon}
\format{An object of class \code{SpatialPixelsDataFrame} with 1673100 rows and 1 columns.}
\usage{
Icon_Primary_Production
}
\description{
Raster data to display the Primary Production Icon
}
\keyword{datasets}
