% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_RData.R
\docType{data}
\name{Icon_Protected_Species}
\alias{Icon_Protected_Species}
\title{Raster data to display the Protected Species Icon}
\format{
An object of class \code{SpatialPixelsDataFrame} with 17408 rows and 1 columns.
}
\usage{
Icon_Protected_Species
}
\description{
Raster data to display the Protected Species Icon
}
\keyword{datasets}
