% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_RData.R
\docType{data}
\name{Icon_Protected_Species_and_Tourism}
\alias{Icon_Protected_Species_and_Tourism}
\title{Raster data to display the Protected Species and Tourism Icon}
\format{
An object of class \code{SpatialPixelsDataFrame} with 18358 rows and 1 columns.
}
\usage{
Icon_Protected_Species_and_Tourism
}
\description{
Raster data to display the Protected Species and Tourism Icon
}
\keyword{datasets}
