% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_RData.R
\docType{data}
\name{Icon_Tourism}
\alias{Icon_Tourism}
\title{Raster data to display the Tourism Icon}
\format{An object of class \code{SpatialPixelsDataFrame} with 1708224 rows and 1 columns.}
\usage{
Icon_Tourism
}
\description{
Raster data to display the Tourism Icon
}
\keyword{datasets}
