% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_RData.R
\docType{data}
\name{Icon_Tuna}
\alias{Icon_Tuna}
\title{Raster data to display the Tuna Icon}
\format{An object of class \code{SpatialPixelsDataFrame} with 1675128 rows and 1 columns.}
\usage{
Icon_Tuna
}
\description{
Raster data to display the Tuna Icon
}
\keyword{datasets}
