% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_RData.R
\docType{data}
\name{Icon_Tuna_Fishery}
\alias{Icon_Tuna_Fishery}
\title{Raster data to display the Tuna Fishery Icon}
\format{An object of class \code{SpatialPixelsDataFrame} with 1697712 rows and 1 columns.}
\usage{
Icon_Tuna_Fishery
}
\description{
Raster data to display the Tuna Fishery Icon
}
\keyword{datasets}
