% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralDecisionTableFunction.R
\name{MakeDecisionTable}
\alias{MakeDecisionTable}
\title{Produce generalized decision tables with provided confidence intervals (optional) on bars.}
\usage{
MakeDecisionTable(Data, BestPerformanceVector, SummaryBestPerformance = NULL,
  OutputDirectory = getwd(), OutputFileName = "DecisionTable",
  GraphicTitle = "Title", RowHeader = "RowHeader",
  ColumnHeader = "ColumnHeader", IconList = NULL, IconColor = "black",
  SummaryRowOption = "Off", ImageWidth = 500, ImageHeight = 800,
  Resolution = 1, GraphicCellHeight = NULL, GraphicCellWidths = NULL,
  BarWidth = 1, BarColors = NULL, IncludeCI = "FALSE",
  Data_UpperCI = NULL, Data_LowerCI = NULL)
}
\arguments{
\item{Data}{A matrix with row and column names (must me matrix even if only 1 row or column), no default}

\item{BestPerformanceVector}{A vector equal in length to the number of rows in Data containing "High" or "Low", no default
"High" means highest value in row will be colored to represent best performance
"Low" means lowest value in row will be colored to represent best performance}

\item{SummaryBestPerformance}{A string containing "High" or "Low" to indicate best performance for summary row, no default.
Only required if "MeanValue", "SumValue", or "MedianValue" options chosen. For other options higher summarized rank correspond with better performance}

\item{OutputDirectory}{A string containing the full path name of folder where resulting graphic should be stored, default = current working directory}

\item{OutputFileName}{A string containing the file name for the resulting graphic, default = "DecisionTable"}

\item{GraphicTitle}{A string containing the title for the decision table graphic, default = "Title"}

\item{RowHeader}{A string containing the descriptive title for row names, default = "RowHeader"}

\item{ColumnHeader}{A string containing the descriptive title for column names, default = "ColumnHeader"}

\item{IconList}{A vector of icon names to be printed right of the GraphicTitle, length may not exceed 3, default = no icons.
A custom black-and-white icon may be printed by providing the filename with .png extension, or an icon supported by the package may be chosen by providing one of the following icon names:
"Commercial_Fisheries_Herring_Mackerel_Lobster",  "Environmental_Considerations",  "Groundfish_Fishery",  "Groundfish_Species",  "Herring_Fishery_Option1",  "Herring_Fishery_Option2",
"Lobster",  "Lobster_Fishery",  "Predator_Fisheries_Tuna_Haddock_Flatfish",  "Predator_Species_Tuna_Haddock_Flatfish",  "Primary_Production",  "Protected_Species",  "Protected_Species_and_Tourism",
"Tourism",  "Tuna",  "Tuna_Fishery"}

\item{IconColor}{A string specifying the color of icons to be printed, only necessary if IconList provide. default = "black"}

\item{SummaryRowOption}{A string indicating summary row option, default = "Off" (no summary row)

     "Off"          = No summary row
     
     "MeanRank"     = Summary row added, values represent mean rank across all rows in each column, highest rank correspond with best performance
     
     "SumRank"      = Summary row added, values represent summed rank across all rows in each column, highest rank correspond with best performance
     
     "MedianRank"   = Summary row added, values represent median rank across all rows in each column, highest rank correspond with best performance 
     
     "MeanValue"    = Summary row added, values represent mean value of Data across all rows in each column, high/low value corresponding with best performance must be specified
     
     "SumValue"     = Summary row added, values represent sum of Data values across all rows in each column, high/low value corresponding with best performance must be specified
     
     "MedianValue"  = Summary row added, values represent median value of Data across all rows in each column, high/low value corresponding with best performance must be specified}

\item{ImageWidth}{A number specifying total graphic image width in number of pixels, default = 500}

\item{ImageHeight}{A number specifying total graphic image height in number of pixels, default = 800}

\item{Resolution}{A number between 1 and 3, default = 1.
This parameter increases the image resolution by scaling the number of pixels in the length and width by an equal amount. 
Increasing the resolution (higher numbers) will result in a larger image which can be adjusted in any image processing software to the desired dimension.}

\item{GraphicCellHeight}{A number specifying height of graphic rows containing data}

\item{GraphicCellWidths}{A vector of length 2 specifying first the width of row label column second data column width}

\item{BarWidth}{A number specifying the width of the plotted bar}

\item{BarColors}{A string specifying single color or a vector of colors equal in length to the number of columns in Data. Default uses colors specified by data dimenstions to visually show ranked performance.}

\item{IncludeCI}{A string specifying the inclusion of confidence levels in plot, if = "TRUE" then Data_UpperCI and Data_LowerCI must also be provided. default = "FALSE"}

\item{Data_UpperCI}{A matrix containing upper confidence levels with row and column names that match Data (must me matrix even if only 1 row or column), no default.}

\item{Data_LowerCI}{A matrix containing lower confidence levels with row and column names that match Data (must me matrix even if only 1 row or column), no default.}
}
\value{
Customized decision table image (.png) with optional confidence intervals on bars.
}
\description{
This function produces a decision table with between 1 and 20 rows 
and columns based on the dimensions of the provided data matrix.
Decision table labels, colors, performance ranking, and decision 
table dimensions can be specified. Confidence intervals may be specified using 
the upper bounds provided in Data_UpperCI and lower bounds provided in
Data_LowerCI matrices (confidence intervals are not automatically generated).
Up to three icons may be printed next to the title, and a summary row can 
also be automatically calculated.
}
\examples{
Produce example decision table and save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = TestMatrix_Ncol_4_Nrow_3, BestPerformanceVector = c("High","High","High","High"))
MakeDecisionTable(Data = TestMatrix_Ncol_4_Nrow_3, BestPerformanceVector = c("High","High","High","High"),
                  IconList = c("Tuna", "Groundfish_Fishery", "Environmental_Considerations"))
MakeDecisionTable(Data = TestMatrix_Ncol_4_Nrow_3, BestPerformanceVector = c("High","High","High","High"),
                  IconList = c("Protected_Species", "Commercial_Fisheries_Herring_Mackerel_Lobster", "Predator_Species_Tuna_Haddock_Flatfish"), 
                  IconColor = "blue")
MakeDecisionTable(Data = TestMatrix_Ncol_4_Nrow_3, BestPerformanceVector = c("High","High","High","High"), 
                  SummaryBestPerformance = c("High","High","High","High"),
                  SummaryRowOption = "SumRank")

}
