% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_TestMatrix_Ncol_12_Nrow_4.R
\docType{data}
\name{TestMatrix_Ncol_12_Nrow_4}
\alias{TestMatrix_Ncol_12_Nrow_4}
\title{TestMatrix_Ncol_12_Nrow_4.RData}
\format{An object of class \code{matrix} with 4 rows and 12 columns.}
\usage{
data(TestMatrix_Ncol_12_Nrow_4)
}
\description{
Example dataset with 12  columns and 4 rows which test default color ranking.
}
\examples{
Produce decision table with same dimensions as example dataset. Save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = TestMatrix_Ncol_12_Nrow_4, BestPerformanceVector = c("High","High","High","High","High","High","High","High","High","High","High","High"))

}
\keyword{datasets}
