% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_TestMatrix_Ncol_16_Nrow_12.R
\docType{data}
\name{TestMatrix_Ncol_16_Nrow_12}
\alias{TestMatrix_Ncol_16_Nrow_12}
\title{TestMatrix_Ncol_16_Nrow_12.RData}
\format{An object of class \code{matrix} with 12 rows and 16 columns.}
\usage{
data(TestMatrix_Ncol_16_Nrow_12)
}
\description{
Example dataset with 16  columns and 12 rows which test default color ranking.
}
\examples{
Produce decision table with same dimensions as example dataset. Save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = TestMatrix_Ncol_16_Nrow_12, BestPerformanceVector = c("High","High","High","High","High","High","High","High","High","High","High","High","High","High","High","High"))

}
\keyword{datasets}
