% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_TestMatrix_Ncol_18_Nrow_2.R
\docType{data}
\name{TestMatrix_Ncol_18_Nrow_2}
\alias{TestMatrix_Ncol_18_Nrow_2}
\title{TestMatrix_Ncol_18_Nrow_2.RData}
\format{An object of class \code{matrix} with 2 rows and 18 columns.}
\usage{
data(TestMatrix_Ncol_18_Nrow_2)
}
\description{
Example dataset with 18  columns and 2 rows which test default color ranking.
}
\examples{
Produce decision table with same dimensions as example dataset. Save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = TestMatrix_Ncol_18_Nrow_2, BestPerformanceVector = c("High","High","High","High","High","High","High","High","High","High","High","High","High","High","High","High","High","High"))

}
\keyword{datasets}
