% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_TestMatrix_Ncol_1_Nrow_16.R
\docType{data}
\name{TestMatrix_Ncol_1_Nrow_16}
\alias{TestMatrix_Ncol_1_Nrow_16}
\title{TestMatrix_Ncol_1_Nrow_16.RData}
\format{An object of class \code{matrix} with 16 rows and 1 columns.}
\usage{
data(TestMatrix_Ncol_1_Nrow_16)
}
\description{
Example dataset with 1  columns and 16 rows which test default color ranking.
}
\examples{
Produce decision table with same dimensions as example dataset. Save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = TestMatrix_Ncol_1_Nrow_16, BestPerformanceVector = c("High","High","High"))

}
\keyword{datasets}
