% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_TestMatrix_Ncol_1_Nrow_20.R
\docType{data}
\name{TestMatrix_Ncol_1_Nrow_20}
\alias{TestMatrix_Ncol_1_Nrow_20}
\title{TestMatrix_Ncol_1_Nrow_20.RData}
\format{An object of class \code{matrix} with 20 rows and 1 columns.}
\usage{
data(TestMatrix_Ncol_1_Nrow_20)
}
\description{
Example dataset with 1  columns and 20 rows which test default color ranking.
}
\examples{
Produce decision table with same dimensions as example dataset. Save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = TestMatrix_Ncol_1_Nrow_20, BestPerformanceVector = c("High","High","High"))

}
\keyword{datasets}
