% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_TestMatrix_Ncol_20_Nrow_1.R
\docType{data}
\name{TestMatrix_Ncol_20_Nrow_1}
\alias{TestMatrix_Ncol_20_Nrow_1}
\title{TestMatrix_Ncol_20_Nrow_1.RData}
\format{An object of class \code{matrix} with 1 rows and 20 columns.}
\usage{
data(TestMatrix_Ncol_20_Nrow_1)
}
\description{
Example dataset with 20  columns and 1 rows which test default color ranking.
}
\examples{
Produce decision table with same dimensions as example dataset. Save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = TestMatrix_Ncol_20_Nrow_1, BestPerformanceVector = c("High","High","High","High","High","High","High","High","High","High","High","High","High","High","High","High","High","High","High","High"))

}
\keyword{datasets}
