% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_TestMatrix_Ncol_2_Nrow_8.R
\docType{data}
\name{TestMatrix_Ncol_2_Nrow_8}
\alias{TestMatrix_Ncol_2_Nrow_8}
\title{TestMatrix_Ncol_2_Nrow_8.RData}
\format{An object of class \code{matrix} with 8 rows and 2 columns.}
\usage{
data(TestMatrix_Ncol_2_Nrow_8)
}
\description{
Example dataset with 2  columns and 8 rows which test default color ranking.
}
\examples{
Produce decision table with same dimensions as example dataset. Save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = TestMatrix_Ncol_2_Nrow_8, BestPerformanceVector = c("High","High"))

}
\keyword{datasets}
