% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_TestMatrix_Ncol_2_Nrow_9.R
\docType{data}
\name{TestMatrix_Ncol_2_Nrow_9}
\alias{TestMatrix_Ncol_2_Nrow_9}
\title{TestMatrix_Ncol_2_Nrow_9.RData}
\format{An object of class \code{matrix} with 9 rows and 2 columns.}
\usage{
data(TestMatrix_Ncol_2_Nrow_9)
}
\description{
Example dataset with 2  columns and 9 rows which test default color ranking.
}
\examples{
Produce decision table with same dimensions as example dataset. Save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = TestMatrix_Ncol_2_Nrow_9, BestPerformanceVector = c("High","High"))

}
\keyword{datasets}
