% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_TestMatrix_Ncol_5_Nrow_17.R
\docType{data}
\name{TestMatrix_Ncol_5_Nrow_17}
\alias{TestMatrix_Ncol_5_Nrow_17}
\title{TestMatrix_Ncol_5_Nrow_17.RData}
\format{An object of class \code{matrix} with 17 rows and 5 columns.}
\usage{
data(TestMatrix_Ncol_5_Nrow_17)
}
\description{
Example dataset with 5  columns and 17 rows which test default color ranking.
}
\examples{
Produce decision table with same dimensions as example dataset. Save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = TestMatrix_Ncol_5_Nrow_17, BestPerformanceVector = c("High","High","High","High","High"))

}
\keyword{datasets}
