% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_TestMatrix_Ncol_5_Nrow_19.R
\docType{data}
\name{TestMatrix_Ncol_5_Nrow_19}
\alias{TestMatrix_Ncol_5_Nrow_19}
\title{TestMatrix_Ncol_5_Nrow_19.RData}
\format{An object of class \code{matrix} with 19 rows and 5 columns.}
\usage{
data(TestMatrix_Ncol_5_Nrow_19)
}
\description{
Example dataset with 5  columns and 19 rows which test default color ranking.
}
\examples{
Produce decision table with same dimensions as example dataset. Save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = TestMatrix_Ncol_5_Nrow_19, BestPerformanceVector = c("High","High","High","High","High"))

}
\keyword{datasets}
