% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_TestMatrix_Ncol_8_Nrow_13.R
\docType{data}
\name{TestMatrix_Ncol_8_Nrow_13}
\alias{TestMatrix_Ncol_8_Nrow_13}
\title{TestMatrix_Ncol_8_Nrow_13.RData}
\format{An object of class \code{matrix} with 13 rows and 8 columns.}
\usage{
data(TestMatrix_Ncol_8_Nrow_13)
}
\description{
Example dataset with 8  columns and 13 rows which test default color ranking.
}
\examples{
Produce decision table with same dimensions as example dataset. Save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = TestMatrix_Ncol_8_Nrow_13, BestPerformanceVector = c("High","High","High","High","High","High","High","High"))

}
\keyword{datasets}
