% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document_TestMatrix_Ncol_9_Nrow_15.R
\docType{data}
\name{TestMatrix_Ncol_9_Nrow_15}
\alias{TestMatrix_Ncol_9_Nrow_15}
\title{TestMatrix_Ncol_9_Nrow_15.RData}
\format{An object of class \code{matrix} with 15 rows and 9 columns.}
\usage{
data(TestMatrix_Ncol_9_Nrow_15)
}
\description{
Example dataset with 9  columns and 15 rows which test default color ranking.
}
\examples{
Produce decision table with same dimensions as example dataset. Save as DecisionTable.png in current working directory.
MakeDecisionTable(Data = TestMatrix_Ncol_9_Nrow_15, BestPerformanceVector = c("High","High","High","High","High","High","High","High","High"))

}
\keyword{datasets}
