% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralDecisionTableFunction.R
\name{MakeDecisionTable}
\alias{MakeDecisionTable}
\title{Produce generalized decision tables.}
\usage{
MakeDecisionTable(Data = NULL, BestPerformanceVector = NULL,
  SummaryBestPerformance = NULL, OutputDirectory = NULL,
  OutputFileName = NULL, GraphicTitle = NULL, RowHeader = NULL,
  ColumnHeader = NULL, IconList = NULL, SummaryRowOption = "Off",
  ImageWidth = 500, ImageHeight = 800, Resolution = 1,
  GraphicCellHeight = NULL, GraphicCellWidths = NULL, BarWidth = 1,
  BarColors = NULL, BarFigureMargins = c(0, 0, 0, 0))
}
\arguments{
\item{Data}{A matrix with row and column names (must me matrix even if only 1 row or column), no default}

\item{BestPerformanceVector}{A vector equal in length to the number of rows in Data containing "High" or "Low", no default
"High" means highest value in row will be colored to represent best performance
"Low" means lowest value in row will be colored to represent best performance}

\item{SummaryBestPerformance}{A string containing "High" or "Low" to indicate best performance for summary row, no default
Only required if "MeanValue", "SumValue", or "MedianValue" options chosen. For other options higher summarized rank correspond with better performance}

\item{OutputDirectory}{A string containing the full path name of folder where resulting graphic should be stored, no default}

\item{OutputFileName}{A string containing the file name for the resulting graphic, no default}

\item{GraphicTitle}{A string containing the title for the decision table graphic, no default}

\item{RowHeader}{A string containing the descriptive title for row names, no default}

\item{ColumnHeader}{A string containing the descriptive title for column names, no default}

\item{IconList}{A vector of icon names to be printed right of the GraphicTitle, length may not exceed 3, default = no icons}

\item{SummaryRowOption}{A string indicating summary row option, default = no summary row
"Off"          No summary row
"MeanRank"     Summary row added, values represent mean rank across all rows in each column, highest rank correspond with best performance
"SumRank"      Summary row added, values represent summed rank across all rows in each column, highest rank correspond with best performance
"MedianRank"   Summary row added, values represent median rank across all rows in each column, highest rank correspond with best performance 
"MeanValue"    Summary row added, values represent mean value of Data across all rows in each column, high/low value corresponding with best performance must be specified
"SumValue"     Summary row added, values represent sum of Data values across all rows in each column, high/low value corresponding with best performance must be specified
"MedianValue"  Summary row added, values represent median value of Data across all rows in each column, high/low value corresponding with best performance must be specified}

\item{ImageWidth}{A number specifying total graphic image width in number of pixels, default = 500}

\item{ImageHeight}{A number specifying total graphic image height in number of pixels, default = 800}

\item{Resolution}{A number between 1 and 3, default = 1.
This parameter increases the image resolution by scaling the number of pixels in the length and width by an equal amount. 
Increasing the resolution (higher numbers) will result in a larger image which can be adjusted in any image processing software to the desired dimension.}

\item{GraphicCellHeight}{A number specifying height of graphic rows containing data}

\item{GraphicCellWidths}{A vector of length 2 specifying first the width of row label column second data column width}

\item{BarWidth}{A number specifying the width of the plotted bar}

\item{BarColors}{A string specifying single color or a vector of colors equal in length to the number of columns in Data. Colors used to visually show rank, defaults specified by dimensions of Data.}

\item{BarFigureMargins}{A vector of length 4 specifying margins of barplots in each graphic cell, default = c(0,0,0,0).
If figure margin error generated when producing decision table, try setting this parameter to the default option, reproducing the graphic, and then specifying parameter as desired}
}
\value{
Customized decision table image (.png)
}
\description{
This function produces a decision table with between 1 and 20 rows 
and columns based on the dimensions of the provided data matrix.
Decision table labels, colors, performance ranking, and decision 
table dimensions can be specified. Up to three icons may be printed 
next to the title, and a summary row can also be automatically calculated.
}
