% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralDecisionTableFunction.R
\name{makeDecisionTable}
\alias{makeDecisionTable}
\title{Produce generalized decision tables with provided confidence intervals (optional) on bars.}
\usage{
makeDecisionTable(
  data,
  rownames = NULL,
  colnames = NULL,
  OutputDirectory = getwd(),
  OutputFileName = "DecisionTable",
  GraphicTitle = "Title",
  RowHeader = "Row_Header",
  ColumnHeader = "Column_Header",
  figureWidth = 500,
  figureHeight = 800,
  resolution = 1,
  graphicCellWidths = NULL,
  graphicCellHeight = NULL,
  barWidth = 1,
  yscalerow = "FALSE",
  SummaryRowOption = "Off",
  visualRank = "FALSE",
  BestPerformanceVector = NULL,
  SummaryBestPerformance = NULL,
  barColors = NULL,
  IncludeCI = "FALSE",
  Data_UpperCI = NULL,
  Data_LowerCI = NULL,
  IconList = NULL,
  IconColor = "black",
  ...
)
}
\arguments{
\item{data}{A data matrix with dimensions matching the desired decision table or a vector equal in length to the number of cells in the desired decision table. By default vectors will fill the table by column according to the dimensions specified by `nrow` and `ncol` arguments, no default.}

\item{rownames}{A vector of strings equal in length to the number of rows to be used as row names for the decision table when data is provided as a vector, no default.}

\item{colnames}{A vector of strings equal in length to the number of columns to be used as column names for the decision table when data is provided as a vector, no default.}

\item{OutputDirectory}{A string containing the full path name of folder where resulting graphic should be stored, default = current working directory}

\item{OutputFileName}{A string containing the file name for the resulting graphic, default = "DecisionTable"}

\item{GraphicTitle}{A string containing the title for the decision table graphic, default = "Title"}

\item{RowHeader}{A string containing the descriptive title for row names, default = "RowHeader"}

\item{ColumnHeader}{A string containing the descriptive title for column names, default = "ColumnHeader"}

\item{figureWidth}{A number specifying total graphic image width in number of pixels, default = 500}

\item{figureHeight}{A number specifying total graphic image height in number of pixels, default = 800}

\item{resolution}{A number between 1 and 3, default = 1.
This parameter increases the image resolution by scaling the number of pixels in the length and width by an equal amount.
Increasing the resolution (higher numbers) will result in a larger image which can be adjusted in any image processing software to the desired dimension.}

\item{graphicCellWidths}{A vector of length 2 specifying first the width of row label column second data column width}

\item{graphicCellHeight}{A number specifying height of graphic rows containing data}

\item{barWidth}{A number specifying the width of the plotted bar}

\item{yscalerow}{A string specifying the type of scaling for the y-axis of bar charts, if = "TRUE" the y-axis scale is chosen independently for each row, else the scale is chosen for the entire decision table, default = "FALSE"}

\item{SummaryRowOption}{A string indicating summary row option, default = "Off" (no summary row)

     "Off"          = No summary row

     "MeanRank"     = Summary row added, values represent mean rank across all rows in each column, highest rank correspond with best performance

     "SumRank"      = Summary row added, values represent summed rank across all rows in each column, highest rank correspond with best performance

     "MedianRank"   = Summary row added, values represent median rank across all rows in each column, highest rank correspond with best performance

     "MeanValue"    = Summary row added, values represent mean value of Data across all rows in each column, high/low value corresponding with best performance must be specified

     "SumValue"     = Summary row added, values represent sum of Data values across all rows in each column, high/low value corresponding with best performance must be specified

     "MedianValue"  = Summary row added, values represent median value of Data across all rows in each column, high/low value corresponding with best performance must be specified
     "WhiskerPlot"  = Summary row added, box and whisker plot summarizes data contained in each column, high/low value corresponding with best performance must be specified}

\item{visualRank}{A string specifying the use of color to visually show ranked relative performance, if = "TRUE" then `BestPerformanceVector` and `SummaryBestPerformance` (if summary row included) must be provided and the `barColors` argument should be set to "defaultRankColor" or provided a vector of unique colors equal in lenght to the number of columns. Default = "FALSE"}

\item{BestPerformanceVector}{A vector equal in length to the number of rows in Data containing "High" or "Low". Default = NULL.
"High" = highest value in row will be colored to represent best performance
"Low" = lowest value in row will be colored to represent best performance}

\item{SummaryBestPerformance}{A string containing "High" or "Low" to indicate best performance for summary row, no default.
Only required if "MeanValue", "SumValue", "MedianValue", or "WhiskerPlot" options chosen. For other options higher summarized rank correspond with better performance}

\item{barColors}{A string specifying single color OR a vector of colors equal in length to the number of data columns, OR "defaultRankColor" which uses default colors to denote ranked performance. Default = grey columns.}

\item{IncludeCI}{A string specifying the inclusion of confidence levels in plot, if = "TRUE" then Data_UpperCI and Data_LowerCI must also be provided. Default = "FALSE"}

\item{Data_UpperCI}{A matrix containing upper confidence levels with row and column names that match Data (must me matrix even if only 1 row or column), no default.}

\item{Data_LowerCI}{A matrix containing lower confidence levels with row and column names that match Data (must me matrix even if only 1 row or column), no default.}

\item{IconList}{A vector of icons (identified by string names) to be printed right of the GraphicTitle, length may not exceed 3, default = no icons.
A custom black-and-white icon may be printed by providing the filename with .png extension, or an icon supported by the package may be chosen by providing one of the following icon names:
"Commercial_Fisheries_Herring_Mackerel_Lobster",  "Environmental_Considerations",  "Groundfish_Fishery",  "Groundfish_Species",  "Herring_Fishery_Option1",  "Herring_Fishery_Option2",
"Lobster",  "Lobster_Fishery",  "Predator_Fisheries_Tuna_Haddock_Flatfish",  "Predator_Species_Tuna_Haddock_Flatfish",  "Primary_Production",  "Protected_Species",  "Protected_Species_and_Tourism",
"Tourism",  "Tuna",  "Tuna_Fishery"}

\item{IconColor}{A string specifying the color of icons to be printed, only necessary if IconList provide. default = "black"}
}
\value{
Customized decision table image (.png) with optional confidence intervals on bars.
}
\description{
This function produces a decision table with dimensions based on those
of the provided data matrix or the specified dimensions if data is provided as vector.
Labels, figure layout and dimensions, and coloring scheme,
can be customized. Confidence intervals may be specified using
the upper bounds provided in Data_UpperCI and lower bounds provided in
Data_LowerCI data inputs (confidence intervals are not automatically generated).
Up to three icons may be printed next to the title, and a summary information can
also be automatically calculated and appended as the last row of the decision table.
}
\examples{
Produce example decision table and save as DecisionTable.png in current working directory.

Example dataframe
set.seed(1)
data_df <- matrix(c(abs(rnorm(30,20,sd=5))), ncol = 5, nrow = 6)
colnames(data_df) <- c("Column 1", "Column 2", "Column 3", "Column 4", "Column 5")
rownames(data_df) <- c("Row 1", "Row 2", "Row 3", "Row 4", "Row 5", "Row 6")

# Decision table with efault graphic settings, custom title and headers
makeDecisionTable(data = data_df, OutputFileName = "Example1", GraphicTitle = "Example decision table with default formatting", RowHeader = "Row header \n describes type \n of data in rows", ColumnHeader = "Column header describes type of data in columns")
# Decision table with ranked coloring scheme, darker colors correspond with better (larger) values in each row
makeDecisionTable(data = data_df, OutputFileName = "Example2", GraphicTitle = "Example decision table with ranked performance", RowHeader = "Row header \n describes type \n of data in rows", ColumnHeader = "Column header describes type of data in columns", BestPerformanceVector = rep("High", nrow(data_df)), barColors = "defaultRankColor", visualRank = "TRUE")
# Decision table with summary row showing sum values for each column, single custom color.
makeDecisionTable(data = data_df, OutputFileName = "Example3", GraphicTitle = "Example decision table with summary row", RowHeader = "Row header \n describes type \n of data in rows", ColumnHeader = "Column header describes type of data in columns", barColors = "cadetblue3", SummaryRowOption = "SumValue", SummaryBestPerformance = "High")

}
